refactor(cli): migrate nim.js to TypeScript (#1271)
## Summary
- Convert `bin/lib/nim.js` (259 lines) to `src/lib/nim.ts`
- Typed interfaces: `NimModel`, `GpuDetection`, `NimStatus`
- 4 pure functions + subprocess-heavy I/O functions (docker, nvidia-smi)
- Co-locate tests (16 tests) with updated mock pattern for dist/ paths

Stacked on #1240. 615 CLI tests pass. Coverage ratchet passes.

Relates to #924 (shell consolidation).
